rename Headbox to HeadInfo
