+++ app/frontend/components/crystals/show/crs_view/head_info.jsx
@@ -2,19 +2,25 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import {dateFormat} from 'components/layouts/date_format'
 
-export const Headbox = ({usrName, crsName, creAt, updAt}) => {
+export const HeadInfo = ({usrName, shwName, creAt, updAt}) => {
+  /* showcase に登録していれば 表示 */
+  const shwcaseLink = (name) => {
+    if (name) return <span>ショーケース : {name}</span>
+    return null
+  }
+
   return (
-    <aside className='frgHead'>
+    <aside className='crsHeadInfo'>
       <span>作成者 : {usrName}</span>
-      <span>クリスタル : {crsName}</span>
       <time>作成日 : {dateFormat(creAt)}</time>
       <time>更新日 : {dateFormat(updAt)}</time>
+      {shwcaseLink(shwName)}
     </aside>
   )
 }
-Headbox.propTypes = {
+HeadInfo.propTypes = {
   usrName: PropTypes.string,
-  crsName: PropTypes.string,
+  shwName: PropTypes.string,
   creAt: PropTypes.string,
   updAt: PropTypes.string
 }
