+++ app/frontend/components/fragments/draftjs/frg_view.jsx
@@ -4,7 +4,7 @@ import {Editor, EditorState, convertFromRaw} from 'draft-js'
 import {CancelContext} from 'components/layouts/app/context'
 import {axiosRails, canceller} from 'components/layouts/axios/instances'
 import {cancelLine} from 'components/layouts/axios/then_catch_funcs'
-import {Headbox} from 'components/fragments/draftjs/frg_view/headbox'
+import {HeadInfo} from 'components/fragments/draftjs/frg_view/head_info'
 import {Actionbox} from 'components/fragments/draftjs/frg_view/actionbox'
 import {Media} from 'components/fragments/draftjs/frg_form/media'
 
@@ -75,7 +75,7 @@ export const FrgView = ({initState}) => {
   return (
     <>
       <h1 className='frgName'>{frgVals.frgName}</h1>
-      <Headbox
+      <HeadInfo
         usrName={addNames.usrName}
         crsName={addNames.crsName}
         creAt={frgVals.creAt}
